requirements.txt
flask>=3.0.0
requests>=2.31.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=1.0.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
rich>=13.9.0
prompt-toolkit>=3.0.48
-e /Users/calico/code/Agent
